refact: rename date_converter to date_extension
